Add a predefined error uri for network failures.
For example, when the dealer is processing an incoming call message
from the caller, the dealer has to send an invocation message to the
callee. If there is a network failure when trying to send the
invocation message we need to send an error response back to the
callee so that it can respond accordingly. Otherwise, the caller
will be blocked waiting for an invocation response which will never
arrive.